redesigned feed items repository

+++ src/main/java/nmd/rss/collector/gae/persistence/GaeFeedItemsRepository.java
@@ -12,10 +12,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import static com.google.appengine.api.datastore.FetchOptions.Builder.withLimit;
-import static com.google.appengine.api.datastore.Query.FilterOperator.EQUAL;
-import static java.lang.Integer.MAX_VALUE;
-import static nmd.rss.collector.gae.persistence.FeedItemEntityConverter.createFeedEntityKind;
 import static nmd.rss.collector.gae.persistence.GaeRootRepository.DATASTORE_SERVICE;
 import static nmd.rss.collector.gae.persistence.GaeRootRepository.getFeedRootKey;
 import static nmd.rss.collector.util.Assert.assertNotNull;
@@ -31,21 +27,22 @@ public class GaeFeedItemsRepository implements FeedItemsRepository {
         assertNotNull(feedId);
         assertNotNull(feedItemsMergeReport);
 
+        final boolean nothingChanged = feedItemsMergeReport.added.isEmpty() && feedItemsMergeReport.removed.isEmpty();
+
+        if (nothingChanged) {
+            return;
+        }
+
+        deleteItems(feedId);
+
+        final List<FeedItem> storedItems = new ArrayList<>();
+        storedItems.addAll(feedItemsMergeReport.retained);
+        storedItems.addAll(feedItemsMergeReport.added);
+
         final Key feedRootKey = getFeedRootKey(feedId);
+        final Entity entity = FeedItemListEntityConverter.convert(feedRootKey, feedId, storedItems);
 
-        for (final FeedItem feedItem : feedItemsMergeReport.removed) {
-            final Key victim = getFeedItemKey(feedId, feedItem.guid);
-
-            if (victim != null) {
-                DATASTORE_SERVICE.delete(victim);
-            }
-        }
-
-        for (final FeedItem feedItem : feedItemsMergeReport.added) {
-            final Entity entity = FeedItemEntityConverter.convert(feedItem, feedRootKey, feedId.toString());
-
-            DATASTORE_SERVICE.put(entity);
-        }
+        DATASTORE_SERVICE.put(entity);
     }
 
     @Override
@@ -58,20 +55,12 @@ public class GaeFeedItemsRepository implements FeedItemsRepository {
             return null;
         }
 
-        final Query query = new Query(createFeedEntityKind(feedId.toString())).setAncestor(feedRootKey);
+        final Query query = new Query(FeedItemListEntityConverter.KIND).setAncestor(feedRootKey);
         final PreparedQuery preparedQuery = DATASTORE_SERVICE.prepare(query);
 
-        final List<Entity> entities = preparedQuery.asList(withLimit(MAX_VALUE));
+        final Entity entity = preparedQuery.asSingleEntity();
 
-        final List<FeedItem> feedItems = new ArrayList<>(entities.size());
-
-        for (final Entity entity : entities) {
-            final FeedItem feedItem = FeedItemEntityConverter.convert(entity);
-
-            feedItems.add(feedItem);
-        }
-
-        return feedItems;
+        return entity == null ? new ArrayList<FeedItem>() : FeedItemListEntityConverter.convert(entity);
     }
 
     @Override
@@ -79,24 +68,14 @@ public class GaeFeedItemsRepository implements FeedItemsRepository {
         assertNotNull(feedId);
 
         final Key feedRootKey = getFeedRootKey(feedId);
-        final Query query = new Query(createFeedEntityKind(feedId.toString())).setAncestor(feedRootKey).setKeysOnly();
+        final Query query = new Query(FeedItemListEntityConverter.KIND).setAncestor(feedRootKey).setKeysOnly();
         final PreparedQuery preparedQuery = DATASTORE_SERVICE.prepare(query);
 
-        final List<Entity> victims = preparedQuery.asList(withLimit(MAX_VALUE));
+        final Entity victim = preparedQuery.asSingleEntity();
 
-        for (final Entity victim : victims) {
+        if (victim != null) {
             DATASTORE_SERVICE.delete(victim.getKey());
         }
-    }
-
-    private Key getFeedItemKey(final UUID feedId, final String itemGuid) {
-        final Key feedRootKey = getFeedRootKey(feedId);
-        final Query query = new Query(createFeedEntityKind(feedId.toString())).setAncestor(feedRootKey).setKeysOnly().setFilter(new Query.FilterPredicate(FeedItemEntityConverter.GUID, EQUAL, itemGuid));
-        final PreparedQuery preparedQuery = DATASTORE_SERVICE.prepare(query);
-
-        final Entity found = preparedQuery.asSingleEntity();
-
-        return found == null ? null : found.getKey();
     }
 
 }
